requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# Data handling and processing
scikit-learn>=1.3.0
scipy>=1.11.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
PyYAML>=6.0

# Hyperparameter tuning
optuna>=3.0.0

# Experiment logging
mlflow>=2.0.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.60.0
